requirements.txt
telebot
requests
google-generativeai
diffusers
torch
pillow